Fix golint issues and move RemoveAllOneFilesystem
