+++ pkg/util/BUILD
@@ -38,7 +38,6 @@ filegroup(
         "//pkg/util/parsers:all-srcs",
         "//pkg/util/pod:all-srcs",
         "//pkg/util/procfs:all-srcs",
-        "//pkg/util/removeall:all-srcs",
         "//pkg/util/resizefs:all-srcs",
         "//pkg/util/rlimit:all-srcs",
         "//pkg/util/selinux:all-srcs",
