--- pkg/util/removeall/BUILD
@@ -1,37 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-load(
-    "@io_bazel_rules_go//go:def.bzl",
-    "go_library",
-    "go_test",
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["removeall_test.go"],
-    embed = [":go_default_library"],
-    deps = [
-        "//pkg/util/mount:go_default_library",
-        "//staging/src/k8s.io/client-go/util/testing:go_default_library",
-    ],
-)
-
-go_library(
-    name = "go_default_library",
-    srcs = ["removeall.go"],
-    importpath = "k8s.io/kubernetes/pkg/util/removeall",
-    deps = ["//pkg/util/mount:go_default_library"],
-)
-
-filegroup(
-    name = "package-srcs",
-    srcs = glob(["**"]),
-    tags = ["automanaged"],
-    visibility = ["//visibility:private"],
-)
-
-filegroup(
-    name = "all-srcs",
-    srcs = [":package-srcs"],
-    tags = ["automanaged"],
-)
